WebKitGTK 2.40.2

+++ Tools/Scripts/webkitpy/common/checkout/scm/scm.py
@@ -97,7 +97,7 @@ class SCM:
         return filenames
 
     def strip_r_from_svn_revision(self, svn_revision):
-        match = re.match("^r(?P<svn_revision>\d+)", unicode(svn_revision))
+        match = re.match(r"^r(?P<svn_revision>\d+)", unicode(svn_revision))
         if (match):
             return match.group('svn_revision')
         return svn_revision
@@ -139,7 +139,7 @@ class SCM:
     def exists(self, path):
         self._subclass_must_implement()
 
-    def changed_files(self, git_commit=None):
+    def changed_files(self, git_commit=None, find_branch=False):
         self._subclass_must_implement()
 
     def changed_files_for_revision(self, revision):
@@ -172,7 +172,7 @@ class SCM:
     def timestamp_of_native_revision(self, path, revision):
         self._subclass_must_implement()
 
-    def create_patch(self, git_commit=None, changed_files=None):
+    def create_patch(self, git_commit=None, changed_files=None, commit_message=False, find_branch=False):
         self._subclass_must_implement()
 
     def committer_email_for_revision(self, revision):
@@ -219,6 +219,8 @@ class SCM:
             if self._filesystem.isdir(filename):
                 if keep_webkitbuild_directory and filename == "WebKitBuild":
                     continue
+                if filename == 'Tools/Scripts/libraries/autoinstalled':
+                    continue
                 self._filesystem.rmtree(filename)
             else:
                 self._filesystem.remove(filename)
@@ -240,6 +242,9 @@ class SCM:
         return len(self.local_commits()) > 0
 
     def discard_local_commits(self):
+        return
+
+    def cleanup_and_optimize_local_repository(self):
         return
 
     def remote_merge_base(self):
